Moved character builder to its own file.
